Some refacroring and Inputsystem implementet
- More consistent naming for components.
- Added an input system. Hardcoded for now, but works. Right click to enable input. If multiple entities has an input component, they will all move.
- Removed unecessary Input class.

+++ mainwindow.h
@@ -25,7 +25,7 @@ public:
     explicit MainWindow(QWidget *parent = nullptr);
     ~MainWindow();
 
-    void showFPS(double deltaTime, double frameCounter);
+    void showFPS(float deltaTime, float frameCounter);
 
     Renderer* getRenderer() { return mRenderer; }
